Reformatted agents code and directory structure
